Decouple BA checkers from model resolution
Pass component classifier context explicitly through BA name, type, legality, consistency, and rules checking so detached strict models do not depend on annex containment. Preserve the legacy normalization and diagnostic behavior in AadlBaModelResolver, then run a read-only strict-model type checker and adapt transition rules to both declarative and strict transitions.
